build_matrix.sh: Fix a shellcheck warning
The warning was:
Use "${var:?}" to ensure this never expands to /* . [SC2115]

(cherry picked from commit 815919247bf40f65777cae4b4eb95f78b4e28917)

+++ build_matrix.sh
@@ -93,7 +93,7 @@ for BUILD_LIBPCAP in ${MATRIX_BUILD_LIBPCAP:-no yes}; do
         done
     else
         echo_magenta 'Use system libpcap'
-        rm -rf "$PREFIX"/*
+        rm -rf "${PREFIX:?}"/*
         make -C ../libpcap distclean || :
         build_tcpdump
     fi
